navigation detail vs new works now, problem: swipe on delete fixen, after using it no new habits are created and it deletes multiple habits upon swiping
plus tasks abhaken können

+++ app/src/main/java/com/example/forma/data/HabitDao.kt
@@ -10,7 +10,7 @@ interface HabitDao {
     suspend fun getAll(): List<Habit>
 
     @Query("SELECT * FROM habits WHERE id = :id")
-    suspend fun getById(id: Int): Habit?
+    suspend fun getById(id: Long): Habit?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(habit: Habit): Long
